refactor(commands/invites): move commands to individual files
